+++ h5rdmtoolbox/wrapper/jsonld.py
@@ -15,9 +15,11 @@ from pydantic import HttpUrl
 from rdflib import Graph, URIRef, BNode, XSD, RDF, SKOS
 from rdflib.plugins.shared.jsonld.context import Context
 
-from h5rdmtoolbox.convention import hdf_ontology
+from h5rdmtoolbox.convention import ontology as hdf_ontology
+from h5rdmtoolbox.convention.ontology.hdf_datatypes import get_datatype
 from .core import Dataset, File
 from .rdf import RDF_TYPE_ATTR_NAME
+from ..convention.ontology.h5ontocls import Datatype
 from ..protocols import H5TbxGroup
 
 _bnode_counter = count()
@@ -126,39 +128,6 @@ def resolve_iri(key_or_iri: str, context: Context) -> str:
     except AttributeError:
         if key_or_iri == 'label':
             return 'http://www.w3.org/2000/01/rdf-schema#label'
-
-
-# def _merge_entries(entries: Dict, clean: bool = True) -> Dict:
-#     _entries = entries.copy()
-#
-#     ids = list(entries.keys())
-#
-#     delete_candidates = []
-#
-#     for _id, entry in entries.items():
-#         for k, v in entry.items():
-#             if clean and len(entry) == 1:
-#                 # remove empty entry, Note, this could be a problem if the entry references elsewhere...
-#                 delete_candidates.append(_id)
-#                 continue
-#             if k not in ('@id', '@type'):
-#                 if isinstance(v, list):
-#                     if all([i in ids for i in v]):
-#                         _entries[_id][k] = [_entries.pop(i) for i in v]
-#
-#                 elif v in ids:
-#                     _entries[_id][k] = _entries.pop(v)
-#     if clean:
-#         for dc in delete_candidates:
-#             _entries.pop(dc, None)
-#     return _entries
-
-
-# def _get_id_from_attr_value(_av, file_url):
-#     if isinstance(_av, (h5py.Dataset, h5py.Group)):
-#         return _get_id(_av, file_url)
-#     else:
-#         return Literal(_av)
 
 
 def _get_id(_node, use_simple_bnode_value: bool, blank_node_iri_base: Optional[Dict]) -> Union[URIRef, BNode]:
@@ -565,7 +534,7 @@ def get_rdflib_graph(source: Union[str, pathlib.Path, h5py.File],
 
     _context = {}
     # if structural:
-    _context['hdf5'] = str(HDF5._NS)
+    _context['hdf5'] = str(HDF5)
     _context.update(context or {})  # = context or {}
 
     assert isinstance(_context, dict)
@@ -675,7 +644,7 @@ def get_rdflib_graph(source: Union[str, pathlib.Path, h5py.File],
             parent_name = obj.parent.name
             parent_node = iri_dict.get(parent_name, None)
             if parent_node is not None:
-                _add_node(g, (parent_node, HDF5.member, obj_node))  # fixme: should not be named obj_node
+                _add_node(g, (parent_node, HDF5.member, obj_node))
 
         if isinstance(obj, h5py.Group):
             if structural:
@@ -719,6 +688,12 @@ def get_rdflib_graph(source: Union[str, pathlib.Path, h5py.File],
                 _add_node(g, (obj_node, HDF5.size, rdflib.Literal(obj.size, datatype=XSD.integer)))
                 _add_node(g, (obj_node, HDF5.dimension, rdflib.Literal(_ndim, datatype=XSD.integer)))
 
+                datatype = get_datatype(obj)
+
+                if datatype:
+                    _add_node(g, (datatype, RDF.type, HDF5.Datatype))
+                    _add_node(g, (obj_node, HDF5.datatype, datatype))
+
                 if obj.dtype.kind == 'S':
                     _add_node(g, (obj_node, HDF5.datatype, rdflib.Literal('H5T_STRING')))
                 elif obj.dtype.kind in ('i', 'u'):
@@ -751,7 +726,14 @@ def get_rdflib_graph(source: Union[str, pathlib.Path, h5py.File],
             logger.debug(f'Create new node for attribute "{ak}": {attr_node}')
 
             if structural:  # add hdf type and name nodes
-                _add_node(g, (attr_node, RDF.type, HDF5.Attribute))
+                if av.dtype.kind == 'S':
+                    _add_node(g, (attr_node, RDF.type, HDF5.StringAttribute))
+                elif av.dtype.kind in ('i', 'u'):
+                    _add_node(g, (attr_node, RDF.type, HDF5.IntegerAttribute))
+                elif av.dtype.kind == 'f':
+                    _add_node(g, (attr_node, RDF.type, HDF5.FloatAttribute))
+                else:
+                    _add_node(g, (attr_node, RDF.type, HDF5.Attribute))
                 attr_def: str = obj.rdf[ak].definition
                 if attr_def:
                     _add_node(g, (attr_node, HDF5.name, rdflib.Literal(ak)))
@@ -1091,36 +1073,14 @@ def dump_file(filename: Union[str, pathlib.Path], skipND) -> str:
         if _id:
             params['id'] = _id
 
-        dtype_dict = {'u': 'H5T_INTEGER',
-                      'i': 'H5T_INTEGER',
-                      'f': 'H5T_FLOAT',
-                      'S': 'H5T_STRING'}
+        datatype = get_datatype(ds)
 
-        # dtype_dict = {'u<8': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_U8LE'},
-        #               'u<16': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_U16LE'},
-        #               'u<32': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_U32LE'},
-        #               'u<64': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_U64LE'},
-        #               'u>8': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_U8BE'},
-        #               'u>16': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_U16BE'},
-        #               'u>32': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_U32BE'},
-        #               'u>64': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_U64BE'},
-        #               'i<8': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_I8LE'},
-        #               'i<16': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_I16LE'},
-        #               'i<32': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_I32LE'},
-        #               'i<64': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_I64LE'},
-        #               'i>8': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_I8BE'},
-        #               'i>16': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_I16BE'},
-        #               'i>32': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_I32BE'},
-        #               'i>64': {'class': 'H5T_INTEGER', 'base': 'H5T_STD_I64BE'},
-        #               'f<32': {'class': 'H5T_FLOAT', 'base': 'H5T_IEEE_F32LE'},
-        #               'f<64': {'class': 'H5T_FLOAT', 'base': 'H5T_IEEE_F64LE'},
-        #               }
-
-        # datatype = dtype_dict.get(f'{ds.dtype.kind}{ds.dtype.byteorder}{ds.dtype.alignment}', None)
-        datatype = dtype_dict.get(ds.dtype.kind, None)
         if datatype:
-            params['datatype'] = datatype
-
+            params['datatype'] = Datatype(id=datatype)
+        else:
+            np_kind = np.dtype(ds.dtype).kind  # see https://numpy.org/doc/stable/reference/generated/numpy.dtype.kind.html
+            if np_kind == "V":
+                params['datatype'] = Datatype(typeClass=HDF5.H5T_COMPOUND)
         ontods = hdf_ontology.Dataset(**params)
 
         if ds.parent.name not in data: